BannerTile->ModBannerTile , implement most the current feedback.
