+++ ExampleMod/Content/Items/Placeable/Banners/ExampleCustomAISlimeNPCBanner.cs
@@ -14,5 +14,4 @@ namespace ExampleMod.Content.Items.Placeable.Banners
 			Item.SetShopValues(ItemRarityColor.Blue1, Item.buyPrice(silver: 10));
 		}
 	}
-	
 }
